test: Geoid format tests
