+++ scripts/GGXF/GGXF.py
@@ -570,7 +570,7 @@ class GridInterpolator:
         # Get the node indices of the corner cells
         nodes = crnr + cellij
         # Multiply the values at the nodes by the interpolation factor
-        nodeprm = data[nodes[:, 1], nodes[:, 0]]
+        nodeprm = data[nodes[:, 0], nodes[:, 1]]
         # Sum the scaled values to get the value at the evaluation point.
         val = (nodef * nodeprm).sum(axis=0)
         return val
