+++ scripts/tests/data/geoid/.gitignore
@@ -1,0 +1,1 @@
+out
