+++ scripts/tests/data/geoid/check/test-ggxf-csv-out.yaml
@@ -1,8 +1,15 @@
 ggxfVersion: GGXF-1.0
-filename: test-out.yaml
+filename: test-ggxf-csv-out.yaml
 content: geoidModel
 title: hybrid geoid
-abstract: "Tiny geoid for testing GGXF implementation"
+abstract: Tiny geoid for testing GGXF implementation
+contentApplicabilityExtent:
+  boundingBox:
+    southBoundLatitude: 22.5
+    westBoundLongitude: -69.0
+    northBoundLatitude: 24.6
+    eastBoundLongitude: -67.0
+  extentDescription: Tiny example geoid.
 interpolationCrsWkt: "GEOGCRS[\"WGS 84 (G2139)\", \n  DYNAMIC[FRAMEEPOCH[2016.0]],\n\
   \  DATUM[\"World Geodetic System 1984 (G2139)\",ELLIPSOID[\"WGS 84\",6378137,298.257223563,LENGTHUNIT[\"\
   metre\",1,ID[\"EPSG\",9001]],ID[\"EPSG\",7030]],ID[\"EPSG\",1309]],\n  CS[ellipsoidal,3,ID[\"\
@@ -21,48 +28,26 @@ targetCrsWkt: "VERTCRS[\"Vertical datum height\",\n VDATUM[\"Imaginary vertical 
   ,ID[\"EPSG\",1124]],\n CS[vertical,1,ID[\"EPSG\",6499]],\n   AXIS[\"Gravity-related\
   \ height (H)\",up],LENGTHUNIT[\"metre\",1,ID[\"EPSG\",9001]]\n ,ID[\"EPSG\",6642]]\
   \    \n"
-operationAccuracy: 0.015
-partyName: "A Hypothetical National Geodetic Survey."
-deliveryPoint: Somewhere
-city: A city
-postalCode: '99999'
-country: United States of America
-contentApplicabilityExtent:
-  boundingBox:
-    southBoundLatitude: 22.5
-    westBoundLongitude: -69.0
-    northBoundLatitude: 24.6
-    eastBoundLongitude: -67.0
 parameters:
 - parameterName: geoidHeight
   sourceCrsAxis: 2
   unitName: metre
   unitSiRatio: 1.0
+operationAccuracy: 0.015
+partyName: A Hypothetical National Geodetic Survey.
+deliveryPoint: Somewhere
+city: A city
+postalCode: '99999'
+country: United States of America
 ggxfGroups:
 - 
-  interpolationMethod: bilinear
   ggxfGroupName: default
+  interpolationMethod: bilinear
   grids:
   - 
-    affineCoeffs:
-    - 22.5
-    - 0.0
-    - 0.7
-    - -69.0
-    - 1.0
-    - 0.0
+    gridName: geoid_grid
+    affineCoeffs: [22.5, 0.0, 0.7, -69.0, 1.0, 0.0]
     iNodeCount: 3
     jNodeCount: 4
-    data: 
-    - - -10.5
-      - -20.5
-      - -30.5
-    - - -11.5
-      - -21.5
-      - -31.5
-    - - -12.5
-      - -22.5
-      - -32.5
-    - - -13.5
-      - -23.5
-      - -33.5
+    data: [[-10.5, -20.5, -30.5], [-11.5, -21.5, -31.5], [-12.5, -22.5,
+        -32.5], [-13.5, -23.5, -33.5]]
