+++ scripts/tests/data/geoid/check/test-inline-csv-out-geoid_grid01.csv
@@ -1,0 +1,13 @@
+nodeLatitude,nodeLongitude,geoidHeight
+22.5,-69,-10.5
+22.5,-68,-20.5
+22.5,-67,-30.5
+23.2,-69,-11.5
+23.2,-68,-21.5
+23.2,-67,-31.5
+23.9,-69,-12.5
+23.9,-68,-22.5
+23.9,-67,-32.5
+24.6,-69,-13.5
+24.6,-68,-23.5
+24.6,-67,-33.5
