+++ scripts/tests/data/geoid/check/test-inline-out.cdl
@@ -1,0 +1,57 @@
+netcdf test-inline-out {
+
+// global attributes:
+		:Conventions = "GGXF-1.0, ACDD-1.3" ;
+		:source_file = "test-inline-out.ggxf" ;
+		:content = "geoidModel" ;
+		:title = "hybrid geoid" ;
+		:summary = "Tiny geoid for testing GGXF implementation" ;
+		:geospatial_lat_min = 22.5 ;
+		:geospatial_lon_min = -69. ;
+		:geospatial_lat_max = 24.6 ;
+		:geospatial_lon_max = -67. ;
+		:extent_description = "Tiny example geoid." ;
+		:interpolationCrsWkt = "GEOGCRS[\"WGS 84 (G2139)\", \n  DYNAMIC[FRAMEEPOCH[2016.0]],\n  DATUM[\"World Geodetic System 1984 (G2139)\",ELLIPSOID[\"WGS 84\",6378137,298.257223563,LENGTHUNIT[\"metre\",1,ID[\"EPSG\",9001]],ID[\"EPSG\",7030]],ID[\"EPSG\",1309]],\n  CS[ellipsoidal,3,ID[\"EPSG\",6423]],\n    AXIS[\"latitude (Lat)\",north,ANGLEUNIT[\"degree\",0.0174532925199433,ID[\"EPSG\",9102]]],\n    AXIS[\"longitude (Lon)\",east,ANGLEUNIT[\"degree\",0.0174532925199433,ID[\"EPSG\",9102]]],\n    AXIS[\"Ellipsoidal height (h)\",up,LENGTHUNIT[\"metre\",1,ID[\"EPSG\",9001]]]\n  ,ID[\"EPSG\",9754]]\n" ;
+		:sourceCrsWkt = "GEOGCRS[\"WGS 84 (G2139)\", \n  DYNAMIC[FRAMEEPOCH[2016.0]],\n  DATUM[\"World Geodetic System 1984 (G2139)\",ELLIPSOID[\"WGS 84\",6378137,298.257223563,LENGTHUNIT[\"metre\",1,ID[\"EPSG\",9001]],ID[\"EPSG\",7030]],ID[\"EPSG\",1309]],\n  CS[ellipsoidal,3,ID[\"EPSG\",6423]],\n    AXIS[\"latitude (Lat)\",north,ANGLEUNIT[\"degree\",0.0174532925199433,ID[\"EPSG\",9102]]],\n    AXIS[\"longitude (Lon)\",east,ANGLEUNIT[\"degree\",0.0174532925199433,ID[\"EPSG\",9102]]],\n    AXIS[\"Ellipsoidal height (h)\",up,LENGTHUNIT[\"metre\",1,ID[\"EPSG\",9001]]]\n  ,ID[\"EPSG\",9754]]\n" ;
+		:targetCrsWkt = "VERTCRS[\"Vertical datum height\",\n VDATUM[\"Imaginary vertical datum\",ID[\"EPSG\",1124]],\n CS[vertical,1,ID[\"EPSG\",6499]],\n   AXIS[\"Gravity-related height (H)\",up],LENGTHUNIT[\"metre\",1,ID[\"EPSG\",9001]]\n ,ID[\"EPSG\",6642]]    \n" ;
+		:parameters.count = 1LL ;
+		:parameters.0.parameterName = "geoidHeight" ;
+		:parameters.0.sourceCrsAxis = 2LL ;
+		:parameters.0.unitName = "metre" ;
+		:parameters.0.unitSiRatio = 1. ;
+		:operationAccuracy = 0.015 ;
+		:institution = "A Hypothetical National Geodetic Survey." ;
+		:deliveryPoint = "Somewhere" ;
+		:city = "A city" ;
+		:postalCode = "99999" ;
+		:country = "United States of America" ;
+		:_NCProperties = "version=2,netcdf=4.9.0,hdf5=1.12.2" ;
+		:_SuperblockVersion = 2 ;
+		:_IsNetcdf4 = 1 ;
+		:_Format = "netCDF-4" ;
+
+group: default {
+
+  // group attributes:
+  		:interpolationMethod = "bilinear" ;
+
+  group: geoid_grid {
+    dimensions:
+    	iNodeCount = 3 ;
+    	jNodeCount = 4 ;
+    variables:
+    	float geoidHeight(iNodeCount, jNodeCount) ;
+    		geoidHeight:_Storage = "contiguous" ;
+    		geoidHeight:_Endianness = "little" ;
+
+    // group attributes:
+    		:affineCoeffs = 22.5, 0., 0.7, -69., 1., 0. ;
+    data:
+
+     geoidHeight =
+  -10.5, -11.5, -12.5, -13.5,
+  -20.5, -21.5, -22.5, -23.5,
+  -30.5, -31.5, -32.5, -33.5 ;
+    } // group geoid_grid
+  } // group default
+}
