+++ scripts/tests/data/geoid/check/test-netcdf-out.yaml
@@ -1,0 +1,52 @@
+ggxfVersion: GGXF-1.0
+filename: test-netcdf-out.yaml
+content: geoidModel
+title: hybrid geoid
+abstract: Tiny geoid for testing GGXF implementation
+interpolationCrsWkt: "GEOGCRS[\"WGS 84 (G2139)\", \n  DYNAMIC[FRAMEEPOCH[2016.0]],\n\
+  \  DATUM[\"World Geodetic System 1984 (G2139)\",ELLIPSOID[\"WGS 84\",6378137,298.257223563,LENGTHUNIT[\"\
+  metre\",1,ID[\"EPSG\",9001]],ID[\"EPSG\",7030]],ID[\"EPSG\",1309]],\n  CS[ellipsoidal,3,ID[\"\
+  EPSG\",6423]],\n    AXIS[\"latitude (Lat)\",north,ANGLEUNIT[\"degree\",0.0174532925199433,ID[\"\
+  EPSG\",9102]]],\n    AXIS[\"longitude (Lon)\",east,ANGLEUNIT[\"degree\",0.0174532925199433,ID[\"\
+  EPSG\",9102]]],\n    AXIS[\"Ellipsoidal height (h)\",up,LENGTHUNIT[\"metre\",1,ID[\"\
+  EPSG\",9001]]]\n  ,ID[\"EPSG\",9754]]\n"
+sourceCrsWkt: "GEOGCRS[\"WGS 84 (G2139)\", \n  DYNAMIC[FRAMEEPOCH[2016.0]],\n  DATUM[\"\
+  World Geodetic System 1984 (G2139)\",ELLIPSOID[\"WGS 84\",6378137,298.257223563,LENGTHUNIT[\"\
+  metre\",1,ID[\"EPSG\",9001]],ID[\"EPSG\",7030]],ID[\"EPSG\",1309]],\n  CS[ellipsoidal,3,ID[\"\
+  EPSG\",6423]],\n    AXIS[\"latitude (Lat)\",north,ANGLEUNIT[\"degree\",0.0174532925199433,ID[\"\
+  EPSG\",9102]]],\n    AXIS[\"longitude (Lon)\",east,ANGLEUNIT[\"degree\",0.0174532925199433,ID[\"\
+  EPSG\",9102]]],\n    AXIS[\"Ellipsoidal height (h)\",up,LENGTHUNIT[\"metre\",1,ID[\"\
+  EPSG\",9001]]]\n  ,ID[\"EPSG\",9754]]\n"
+targetCrsWkt: "VERTCRS[\"Vertical datum height\",\n VDATUM[\"Imaginary vertical datum\"\
+  ,ID[\"EPSG\",1124]],\n CS[vertical,1,ID[\"EPSG\",6499]],\n   AXIS[\"Gravity-related\
+  \ height (H)\",up],LENGTHUNIT[\"metre\",1,ID[\"EPSG\",9001]]\n ,ID[\"EPSG\",6642]]\
+  \    \n"
+operationAccuracy: 0.015
+partyName: A Hypothetical National Geodetic Survey.
+deliveryPoint: Somewhere
+city: A city
+postalCode: '99999'
+country: United States of America
+contentApplicabilityExtent:
+  boundingBox:
+    southBoundLatitude: 22.5
+    westBoundLongitude: -69.0
+    northBoundLatitude: 24.6
+    eastBoundLongitude: -67.0
+parameters:
+- parameterName: geoidHeight
+  sourceCrsAxis: 2
+  unitName: metre
+  unitSiRatio: 1.0
+ggxfGroups:
+- 
+  ggxfGroupName: default
+  interpolationMethod: bilinear
+  grids:
+  - 
+    gridName: geoid_grid
+    affineCoeffs: [22.5, 0.0, 0.7, -69.0, 1.0, 0.0]
+    iNodeCount: 3
+    jNodeCount: 4
+    data: [[-10.5, -20.5, -30.5], [-11.5, -21.5, -31.5], [-12.5, -22.5,
+        -32.5], [-13.5, -23.5, -33.5]]
