+++ scripts/tests/data/geoid/check/test-points-out.csv
@@ -1,0 +1,4 @@
+nodeLatitude,nodeLongitude,geoidHeight
+22.6,-68.9,-11.6429
+23.8,-67.2,-30.3571
+24.5,-67.1,-32.3571
