+++ scripts/tests/data/geoid/test-netcdf.cdl
@@ -1,8 +1,8 @@
-netcdf test_geoid {
+netcdf test-inline-out {
 
 // global attributes:
 		:Conventions = "GGXF-1.0, ACDD-1.3" ;
-		:source_file = "test_geoid.ggxf" ;
+		:source_file = "test-inline-out.ggxf" ;
 		:content = "geoidModel" ;
 		:title = "hybrid geoid" ;
 		:summary = "Tiny geoid for testing GGXF implementation" ;
@@ -40,7 +40,7 @@ group: default {
     	iNodeCount = 3 ;
     	jNodeCount = 4 ;
     variables:
-    	float geoidHeight(jNodeCount, iNodeCount) ;
+    	float geoidHeight(iNodeCount, jNodeCount) ;
     		geoidHeight:_Storage = "contiguous" ;
     		geoidHeight:_Endianness = "little" ;
 
@@ -49,10 +49,9 @@ group: default {
     data:
 
      geoidHeight =
-  -10.5, -20.5, -30.5,
-  -11.5, -21.5, -31.5,
-  -12.5, -22.5, -32.5,
-  -13.5, -23.5, -33.5 ;
+  -10.5, -11.5, -12.5, -13.5,
+  -20.5, -21.5, -22.5, -23.5,
+  -30.5, -31.5, -32.5, -33.5 ;
     } // group geoid_grid
   } // group default
 }
