+++ scripts/tests/data/geoid/test-points.csv
@@ -1,0 +1,4 @@
+nodeLatitude,nodeLongitude
+22.6,-68.9
+23.8,-67.2
+24.5,-67.1
