--- scripts/tests/data/geoidModel/compile_ggxf.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-rm -f test_geoid.ggxf test_geoid.cdl
-python3 ../../../scripts/ggxf.py -v -n write_cdl=true test_geoid.yaml -o test_geoid.ggxf
