--- scripts/tests/data/geoidModel/test.grid.csv
@@ -1,13 +1,0 @@
-X,Y,i,j,geoidHeight
-22.5,-69.0,0,0,-10.5
-22.5,-68.0,1,0,-20.5
-22.5,-67.0,2,0,-30.5
-23.2,-69.0,0,1,-11.5
-23.2,-68.0,1,1,-21.5
-23.2,-67.0,2,1,-31.5
-23.9,-69.0,0,2,-12.5
-23.9,-68.0,1,2,-22.5
-23.9,-67.0,2,2,-32.5
-24.6,-69.0,0,3,-13.5
-24.6,-68.0,1,3,-23.5
-24.6,-67.0,2,3,-33.5
