--- scripts/tests/data/geoidModel/test_points-out.csv.chk
@@ -1,3 +1,0 @@
-x,y,geoidHeight
-22.6,-68.9,-11.6429
-23.8,-67.2,-30.3571
