--- scripts/tests/data/geoidModel/test_points.csv
@@ -1,3 +1,0 @@
-x,y
-22.6,-68.9
-23.8,-67.2
